refactor: calendar-event-rendering を lib/calendar/ 配下へ移動（Phase 3）
lib 直下に取り残されていたカレンダー用ユーティリティを lib/calendar/event-rendering.ts へ移し、他のカレンダーモジュールと配置を揃える。
