+++ lib/calendar/event-rendering.ts
